requirements.txt
Django==2.2.6
django-cors-headers==3.1.1
django-restframework==0.0.1
django-webpack-loader==0.6.0
djangorestframework==3.10.3
graphviz==0.13
PyQt5==5.13.1
PyQt5-sip==12.7.0
pytz==2019.3
requests==2.7.0
selenium==3.141.0
sqlparse==0.3.0
urllib3==1.25.6
